Trust me bro

+++ examples/loop_invariant.m
@@ -2,17 +2,17 @@
 
 setup()
 {
-    int8 n = 37; // Arbitrary value
+    int16 n = 37; // Arbitrary value
 }
 
 mainloop()
 {
-    uint32 i = 1; // For outer loop
+    uint16 i = 1; // For outer loop
 
     while (i <= 10)
     {
         uint16 L = i * (n + 2);
-        uint32 j = 1; // For inner loop
+        uint16 j = 1; // For inner loop
         while (j <= 10)
         { 
             uint16 R = L * 10 * j; // Interesting calculation
